[CSP] Do not match wildcard directives to FTP
The spec [1] says: "This logic means that in order to allow a resource
from a non-HTTP(S) scheme, it has to be either explicitly specified
(e.g. default-src * data: custom-scheme-1: custom-scheme-2:), or the
protected resource must be loaded from the same scheme."

We have logged CspWouldBlockIfWildcardDoesNotMatchFtp in stable and the
number [2] is extremely low, below 0.000001%, so we can safely remove
this behavior.

[1] https://www.w3.org/TR/CSP3/#match-url-to-source-expression
[2] https://uma.googleplex.com/p/chrome/timeline_v2/?sid=5e999f04cc09f20bce48dcf53e20da87

Bug: 1477193
Change-Id: I8886cb81e6dff5e05ee64ea12a55b96e1db588dc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5123339
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Commit-Queue: Jonathan Hao <phao@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1240539}
NOKEYCHECK=True
GitOrigin-RevId: ec7a37187a5a76b23b4a0817b486e580ed5aa087