Fix paint invalidation delay check in LayoutObject::WillRenderImage
Only suspend when paint invalidation is delayed and the object is marked
as needing to check for paint invalidation. When the image that's being
delayed is outside the viewport but end being painted because it's
within the "interest rect" it could otherwise end up in a state where
it is suspended but won't be painted again (lacking other external
stimuli that trigger a paint invalidation).

Fixed: 1513304
Change-Id: I085784f178b38031ac420cabee02183242a6c373
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5149769
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: Fredrik Söderquist <fs@opera.com>
Cr-Commit-Position: refs/heads/main@{#1240564}
NOKEYCHECK=True
GitOrigin-RevId: 0a6fd6fc73dfd72ae171b62e22eb657c70990695