[iOS Blink] Disable Mojo methods related the screen capture
https://crrev.com/c/5116384 defined CapturedSurfaceController
for the Capture Surface Control APIs. But it broke the build
on the ios-blink bot because the iOS blink port didn't enable
the screen capture feature yet. So, to restore the bot, this
CL disables screen capture related Mojo methods of MediaStream
and MediaDevicesDispatcherHost Mojo interfaces on iOS Blink in
order to avoid calling the methods defined in
CapturedSurfaceController.

Bug: 1466247, 1513666
Change-Id: I2eda841d1b4d00af1dd73f4dbb848053699d833e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5144027
Reviewed-by: Guido Urdaneta <guidou@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Commit-Queue: Gyuyoung Kim <gyuyoung@igalia.com>
Cr-Commit-Position: refs/heads/main@{#1240571}
NOKEYCHECK=True
GitOrigin-RevId: 8b39ba585b413b02bdb54f35312e0d2ccf9d9f05