In-effect animations should force a stacking context
An in-effect animation includes those that are in the before or after
phase with the animation having the corresponding fill. Per spec[1],
such animations should have a side-effect of being equivalent to having
`will-change` set to the target property.

https://crrev.com/46005e74c changed opacity animations to no longer have
clamp to a value below exactly 1. It turns out having a non-1 value was
what was causing opacity animations to force a stacking context; the
animation effect was only considered to be affecting a property if the
effect was current.

This CL updates the 'affecting' state to include animations that are in
effect.

[1] https://drafts.csswg.org/web-animations-1/#side-effects-section

Bug: 1511227
Change-Id: I05d7d274d004c55781e25734162eb3cc42ff36bb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5155307
Reviewed-by: Steve Kobes <skobes@chromium.org>
Commit-Queue: David Bokan <bokan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1241421}
NOKEYCHECK=True
GitOrigin-RevId: 5513e11571b0d3418e0dbb86458fc07db5c51133