borderless: Loosen minimum size for window.open / window.resizeTo|By
- Introduces a new constant for specific for borderless apps so that
borderless apps can create smaller windows (at smallest 18x18px).
- Moves minimum-size adjusting for window.open calls to chrome/
instead of blink if the opener_frame is borderless and navigation policy
is a popup.
- Uses borderless-specific minimum-size for window.resizeTo|By in
blink, if they are called in an app with borderless display-mode. The
correct permission is ensured in chrome/ side.

Note: When window.open is called, display-mode for the new frame is
still undefined as the app hasn't loaded yet and thus using the opener
frame.

Bug: b/307182741, b/307160156
Change-Id: Id8b37e930d0aaa4293629a82c48714c3049ae7cc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4999178
Commit-Queue: Sonja Laurila <laurila@google.com>
Reviewed-by: Mike Wasserman <msw@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1242003}
NOKEYCHECK=True
GitOrigin-RevId: 4da654e4c414ef33349f3d7efe01e6ccdc6fce30