Use LayoutObject style for OOFCandidateStyleIterator
The LayoutObject style is the right one to use for layout purposes. It
differs from the Element style in some cases. In particular, the
computed writing-mode of the body is propagated to the LayoutObject
style of the root element, but does not affect the computed value for
the root element.

This caused a DCHECK() failure when the constraint_space writing-mode
differed from the ComputedStyle used during layout.

Bug: 1512988
Change-Id: I0b25bd8a05556cbe84a5324ff8724836336ab1ea
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5139690
Commit-Queue: Rune Lillesveen <futhark@chromium.org>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1242343}
NOKEYCHECK=True
GitOrigin-RevId: 4b4ffe0a7b07499c5c24f329130dfde06c755d67