When testing MPC hits for matching inherited data, compare by-value.
This makes for a more expensive comparison, but we get it back by
getting complete MPC hits more often -- saving both RAM and CPU.
(The effect is especially pronounced on style recalcs, where the
MPC already most likely contains nearly everything we'd want.)
In particular, if two elements get the same state from different
parts in the tree, the pointers are not the same, but the data
is equivalent. E.g.

  * div
    * div class="a"
      * span class="b"
    * div class="a"
      * span class="b"  <-- would get a partial MPC hit before, full now

Note also that there is a knock-on effect from reusing the ComputedStyle
from the MPC hit, in that any similar children of the two spans would
now automatically have the same group pointers for inherited data,
too (so when we get a match, we don't need the by-value comparison for
all the children; the pointers will be the same and the comparison will
short-circuit).

This saves a large amount of ComputedStyle RAM:

  Video            766 →   395 kB ( -371 kB)
  Extension       3002 →  1670 kB (-1332 kB)
  News            1724 →  1442 kB ( -282 kB)
  ECommerce        238 →   198 kB (  -40 kB)
  Social1          755 →   588 kB ( -167 kB)
  Social2          536 →   461 kB (  -75 kB)
  Encyclopedia    1658 →  1293 kB ( -365 kB)
  Sports           541 →   409 kB ( -132 kB)
  Search           415 →   238 kB ( -177 kB)

It would be possible to force-rewrite equivalent pointers into
each other to save yet more RAM (one of them could be GC-ed away
even if we did not have an MPC match), but from a quick prototyping,
the effect is mostly small, and it costs a little more CPU (~1%).

It also saves CPU (style perftest, Zen 2, LTO but no PGO):

  Initial style (µs)     Before     After    Perf      95% CI (BCa)
  =================== ========= ========= ======= =================
  ECommerce                8456      8373   +1.0%  [ +0.4%,  +1.5%]
  Encyclopedia            71749     70657   +1.5%  [ +1.0%,  +2.4%]
  Extension               93541     86623   +8.0%  [ +7.2%,  +8.6%]
  News                    36167     35159   +2.9%  [ +2.4%,  +3.7%]
  Search                  10621     10295   +3.2%  [ +2.7%,  +3.7%]
  Social1                 20412     19897   +2.6%  [ +2.1%,  +3.1%]
  Social2                 13418     13165   +1.9%  [ +1.3%,  +2.6%]
  Sports                  39573     38658   +2.4%  [ +1.3%,  +6.3%]
  Video                   30046     27558   +9.0%  [ +8.5%,  +9.6%]
  Geometric mean                            +3.6%  [ +3.0%,  +4.1%]

  Recalc style (µs)      Before     After    Perf      95% CI (BCa)
  =================== ========= ========= ======= =================
  ECommerce                5267      4418  +19.2%  [+18.1%, +20.5%]
  Encyclopedia            59969     56885   +5.4%  [ +5.0%,  +6.3%]
  Extension               79453     70043  +13.4%  [+12.2%, +14.5%]
  News                    27126     22136  +22.5%  [+21.5%, +23.6%]
  Search                   5500      4340  +26.7%  [+25.4%, +27.8%]
  Social1                 12423     10405  +19.4%  [+18.1%, +20.4%]
  Social2                  9104      8050  +13.1%  [+11.8%, +14.1%]
  Sports                  22436     19062  +17.7%  [+15.0%, +26.9%]
  Video                   17912     12632  +41.8%  [+40.8%, +42.8%]
  Geometric mean                           +19.6%  [+18.5%, +20.8%]

Speedometer3, M1 Pinpoint (lower is better except Score,
significant results at 95% CI only, ordered by increasing goodness):

  TodoMVC-JavaScript-ES6-Webpack            [ +0.6%,  +2.2%]
  TodoMVC-React-Redux                       [ +0.2%,  +0.9%]
  Editor-TipTap                             [ +0.2%,  +0.6%]
  Charts-chartjs                            [ -0.9%,  -0.1%]
  TodoMVC-Backbone                          [ -0.8%,  -0.1%]
  TodoMVC-JavaScript-ES5                    [ -0.8%,  -0.1%]
  TodoMVC-jQuery                            [ -0.8%,  -0.2%]
  TodoMVC-React                             [ -1.1%,  -0.3%]
  TodoMVC-Angular                           [ -1.5%,  -0.3%]
  TodoMVC-Preact                            [ -3.0%,  -1.4%]
  TodoMVC-React-Complex-DOM                 [ -3.7%,  -2.6%]
  TodoMVC-WebComponents                     [ -5.5%,  -4.8%]
  TodoMVC-Lit                               [ -6.6%,  -5.7%]

  Score                                     [ +0.7%,  +1.0%]

We put it behind the CSSMPCImprovement flag (default on), to be able
to fold it into the ongoing GWS experiment.

Bug: 1504480
Change-Id: Id47631629f90ad049181957adce47e330a512693
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5149687
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Steinar H Gunderson <sesse@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1242371}
NOKEYCHECK=True
GitOrigin-RevId: c9c6c875f6b8df9aeba5653c60571143e475ba08