[FedCM] Fix promise resolution in multiIDP array case
Currently, the first IDP in the array will result in the promise being
resolved. This means that choosing an account from the non-first IDP
in a get() call will result in failure even if the token request is
successful. To fix this, we store the list of IDPs along with their
resolver, and we check whether the list contains the selected IDP
when we need to resolve or reject the promise.

Bug: 1515128
Change-Id: Ibd3e4a0b52c2b827b0daa286f3d40987f414b882
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5169965
Commit-Queue: Nicolás Peña <npm@chromium.org>
Reviewed-by: Christian Biesinger <cbiesinger@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1242986}
NOKEYCHECK=True
GitOrigin-RevId: 6d9aea1353216ca43e794fa6cc650731436f6e8f