Add tracing for PointerEvent.deviceId
These tracing points are being added to assist in understanding scenarios in which users are not getting the right deviceId. It can be inconvenient or sometimes impossible to debug or log issues on consumer devices; therefore the traces will help to quickly diagnose the way deviceId is failing on a device.

Bug: 1475303
Change-Id: I44747f9973fbd0a8d90bb36d289775f4649fb614
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5132752
Reviewed-by: Robert Liao <robliao@chromium.org>
Commit-Queue: Sahir Vellani <sahir.vellani@microsoft.com>
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1243030}
NOKEYCHECK=True
GitOrigin-RevId: 28ed386d92f5caca8362165219ff9efe2dec6551