Remove UseLegacyBackgroundSizeShorthandBehavior.
This workaround is marked as “can be removed at any time after 2015”,
and causes NOTREACHED() when background: is set to a custom property.
(This is because we set background to expand to background-size
in the .json5, but the parser does not when this flag is set.)
This fixes one of the top CSS crashes on WebView, since evidently,
some ad out there is doing exactly that.

Bug: 277157
Fixed: 1423568
Change-Id: I20d7581ca425db5b27277f38932325647dbd6bd6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5153781
Reviewed-by: Richard (Torne) Coles <torne@chromium.org>
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Commit-Queue: Steinar H Gunderson <sesse@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1243433}
NOKEYCHECK=True
GitOrigin-RevId: 7609ba48b5df63e9ce4e456c672dbe9aabb82019