view-transition: Fix RuleSet invalidation for VT pseudos
The view-transition pseudo elements are manually style invalidated
whenever the document element is [1]. This assumes that the document
element will be invalidated whenever the view transition pseudo styles
need to be recalculated.

It seems this was often be the case but due to over-invalidation.
https://crrev.com/c3afc7e99450 improved the style system to invalidate
only when a matching selector's rules changed, rather than the whole
stylesheet. As a consequence, changing a ::view-transition rule in a
stylesheet no longer (incidentally) invalidates the document element so
the changes aren't made visually.

RuleSetChange based invalidation is applied by walking the Element
tree[2] which doesn't include pseudo elements. Instead, a pseudo element
selector should match its originating element which will cause its
invalidation. Setting the `dynamic_pseudo` on the result will prevent
the rule from being collected in [3]. The view transition pseudos were
only matching when in the context of a pseudo element which prevented
this invalidation.

[1] https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/renderer/core/dom/node.cc;l=1307;drc=35cedca3b1ad347c2b576e1d8eed30102a625a31
[2] https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/renderer/core/css/style_engine.cc;l=2225;drc=36f47fc1f93ac985138ef4579fdead05ae19c9ed
[3] https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/renderer/core/css/element_rule_collector.cc;l=1178;drc=e5f8ff312fe542f39c7d34b4440f18d1500b15f9

Bug: 1508345
Change-Id: Ia92ecfb2fb5138920eda145f83b5367f9b9ce77d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5147306
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Reviewed-by: Dominic Farolino <dom@chromium.org>
Commit-Queue: David Bokan <bokan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1243529}
NOKEYCHECK=True
GitOrigin-RevId: aa1c0df9bc1b2e6d253451aef7702994899252fe