Move WebTestControlHost to be associated based on RenderFrameHost
Associate the channel with the frame's associated channel. This
gets rid of the WebTestRenderThread interface and everything is sent
via frame bound interfaces.

Since TestRunner no longer has a WebTestControlHost and it must
be accessed via a WebFrameTestProxy we need to pass the source
of the events into various TestRunner interfaces.

Bug: 993189
Change-Id: Ie4629f90d3bad5ae6a1530c2a7bb87b51bf76280
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5105229
Reviewed-by: danakj <danakj@chromium.org>
Commit-Queue: Dave Tapuska <dtapuska@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1243614}
NOKEYCHECK=True
GitOrigin-RevId: 64faea4fd9465ba61a48a5a617f04e59464a6b52