CoopAccessMonitor: Use normal priority task runner for disconnect
CoopAccessMonitor uses TaskType::kInternalHighPriorityLocalFrame
for the task runner associated with its remote. Since it's a remote
interface with no responses, this is only used for disconnect
notifications. We'd like to remove this task type, but using a normal priority task runner causes a WPT test failure:
 1. The Bind() call in InstallCoopAccessMontor will (silently) fail if
    underlying pipe was torn down in the browser process before binding,
    which can happen if overwriting a RFH's `coep_reporter_` in
    TakeNewDocumentPropertiesFromNavigation (which causes disconnects
    when the previous reporter is destroyed). This causes a task to be
    scheduled to run the disconnect callback.
 2. The Bind() happens again with a new monitor, which overwrites the
    old one.
 3. The disconnect (for the old remote) callback runs, removing the new
    entry and causing the test to fail.

This happens because of a subtle task race introduced in
crrev.com/c/5049015. Steps (2) and (3) above are reversed if using a
high priority task runner for the disconnect callback. But the reason
this stopped working with a normal priority task runner is because
previously the old monitor's remote would be reset in step (2). Since
CoopAccessMonitor are now GCed objects, the remote needs to be reset
manually to avoid past entries clobbering newer entries.

Bug: 1426364
Change-Id: If44c1b44f6f52ff6112d160af54ffd4b38b359b1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5133001
Commit-Queue: Scott Haseley <shaseley@chromium.org>
Reviewed-by: David Bokan <bokan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1244367}
NOKEYCHECK=True
GitOrigin-RevId: f2d065cbd3c8a9be0f1d7f86599c4272ec6bbd26