Async clipboard: Omit permissions check inside event handlers
For example, no permission is required when using
navigator.clipboard.read() inside a "paste" event handler.

This behavior is flag gated for now, but since the behavior is
already described in the spec, the new WPT is not marked as
tentative. The WPT is in internal for now but should be moved
to external once testdriver can be updated.

Bug: 1501971

Change-Id: Ia4395a02c30472616b55e2419081f534f44db388
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5018977
Commit-Queue: Evan Stade <estade@chromium.org>
Auto-Submit: Evan Stade <estade@chromium.org>
Reviewed-by: Kent Tamura <tkent@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1244433}
NOKEYCHECK=True
GitOrigin-RevId: d9ecf073152e142f101f58da158b705ac9a3a510