[bfcache] Make unload a blocklisted feature to capture location
This CL adds a new BFCache blocklisted reason "kUnloadHandler". The purpose of this is to use existing functions that capture the source location of blocklisted features for unload handlers. It does not affect the ongoing analysis of unload handler usage because it keeps the original reasons "kUnloadExistsInMain/SubFrame" as they are, and reports the new reason separately.

Bug: 1513120
Change-Id: I0e4b47ad0fb90cf2566640905a80215d86e05ae7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5134730
Reviewed-by: Yuzu Saijo <yuzus@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Mingyu Lei <leimy@chromium.org>
Commit-Queue: Kurumi Muto <mkurumi@google.com>
Cr-Commit-Position: refs/heads/main@{#1244599}
NOKEYCHECK=True
GitOrigin-RevId: b3137e8971403be865053cda3cb0cc0372121d97