Correctly handle a zero value for X25519 deriveBits' length parameter
The Secure Curves spec draft states that we should trunk the derived
bits based on any non-null value passed in the length parameter. Our
currently implementation has been considering zero like null, so we
are returning the entire derived bits array.

The bug 1433707 mentions on ongoing discussion related to the lack of
interoperability on the handling of the deriveBits' length parameter
in several algorithms, being the X25519 one of them. The proposed
change in the spec could be backward-incompatible, so we need time to
evaluate the impact on Chrome's implementation.

While the mentioned spec change discussion continues, this CL fixes a bug in our implementation that considered zero as null. With this change zero will return an empty vector, which matches the current spec.

Bug: 1433707
Change-Id: Ifbf5829db04bbc7c9ff916d27b586feae345b536
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5013612
Commit-Queue: Javier Fernandez <jfernandez@igalia.com>
Reviewed-by: Elly FJ <ellyjones@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1244653}
NOKEYCHECK=True
GitOrigin-RevId: 5b672ce7088999b3d9b09267082a613849e208d5