[EditContext] Report editable InputType when focus is in <canvas>
Selection currently can't be inside a <canvas>, so when an EditContext
is associated with a <canvas> and the user is typing inside it, the
<canvas> has focus but selection will still lie outside.

Since InputMethodController::TextInputType() uses selection position,
this can result in the wrong WebTextInputType being reported, which
can cause the IME to fail to appear.

We should consider allowing selection to exist inside <canvas>
when there is an associated EditContext. However, this may require
spec changes, and the implementation changes look potentially
complicated, requiring edits to both hit testing for replaced
content as well as selection position calculation.

Meanwhile, since this is a pretty significant bug, check for
this case in InputMethodController::TextInputType() to
ensure we return the right result when the user is editing
inside a <canvas>.

Bug: 1513971
Change-Id: I6961fc2dd1168d99b511e652901bb87bc2bab480
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5150014
Commit-Queue: Dan Clark <daniec@microsoft.com>
Reviewed-by: Anupam Snigdha <snianu@microsoft.com>
Reviewed-by: Koji Ishii <kojii@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1244851}
NOKEYCHECK=True
GitOrigin-RevId: 8ac6d3aa40fe8ad1597245450da57b7d9918d6cd