Distinguish click checks in resize-lock-input test.
There are two identical expectations for this test, and the expectation
text indicates a flake. WPT failure messages are not verbose at all, so
it's not clear which expectation is failing. Making them unique helps
us understand if the flake is from click handling vs the actual
resizing, or some race between the two.

R=gtanzer@chromium.org

Bug: 1372536
Change-Id: Icafc2ea3758e06658c4544a1a9789cfd666a1979
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5180718
Reviewed-by: Garrett Tanzer <gtanzer@chromium.org>
Commit-Queue: Andrew Verge <averge@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1244871}
NOKEYCHECK=True
GitOrigin-RevId: 66485caff3b678479c875fbb273367bd9cf71ae6