Fix saving embedded images from pages resulting from form POST.
This requires being able to distinguish between when the image we're
saving is the result of a form POST (being rendered as an image
document), and when it's an image embedded in an HTML page (and was
itself fetched via a normal GET).

To this end, the context menu data now includes whether this is a
document which exists only to host a single image, media or plugin
element. This allows the browser to distinguish whether it's a
subresource that's being interacted with, pass that into WebContents,
and not attempt to use the post data if it is.

Fixed: 1285983
Change-Id: I6df4e1e27631a460cc87be229b67065e68132543
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5181634
Auto-Submit: Jeremy Roman <jbroman@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Commit-Queue: Avi Drissman <avi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1245052}
NOKEYCHECK=True
GitOrigin-RevId: 4d408d8a86ad7863c0cabdc25f30e2aaea4f3164