Use network::ResourceRequest instead of WebURLRequest in ResourceFetcher
blink::WebURLRequest is used for the argument of
URLLoaderThrottleProvider::CreateThrottles() method after [1]. But we
are planing to call this method from the background thread where we
don't use blink::WebURLRequest [2]. Also we are planing to deprecate
blink::WebURLRequest, and planing to use network::ResourceRequest
instead [3].

So this CL changes the argument type of CreateThrottles() from
blink::WebURLRequest to network::ResourceRequest.
A network::ResourceRequest is created in ResourceLoader::Start(). So we
pass the reference of the structure through ResourceFetcher::
CreateURLLoader().

In frames:
  ResourceLoader::Start()
  -> ResourceFetcher::CreateURLLoader()
    -> LoaderFactoryForFrame::CreateURLLoader()
    |-> URLLoaderThrottleProvider::CreateThrottles()
    |-> ServiceWorkerNetworkProviderForFrame::
    |     GetSubresourceLoaderFactory()
    |-> URLLoaderFactory::CreateURLLoader()

In workers:
  ResourceLoader::Start()
  -> ResourceFetcher::CreateURLLoader()
    -> LoaderFactoryForWorker::CreateURLLoader()
    |-> WebWorkerFetchContext::CreateThrottles()
    | -> WebServiceWorkerFetchContextImpl::CreateThrottles() or
    |    DedicatedOrSharedWorkerFetchContextImpl::CreateThrottles()
    |  -> URLLoaderThrottleProvider::CreateThrottles()
    |-> URLLoaderFactory::CreateURLLoader()

mojom::blink::RequestContextType in blink::WebURLRequest is checked in
the code path. But this type member is not in network::ResourceRequest.
network::ResourceRequest has a network::mojom::RequestDestination
member. And some of these checks of mojom::blink::RequestContextType can
be safely converted to network::mojom::RequestDestination without any
behavior change. But the check in ResourceFetcher::CreateURLLoader()
needs the mojom::blink::RequestContextType. Also we need to pass
`service_worker_race_network_request_token` to
LoaderFactoryForWorker::CreateURLLoader(), and
`is_from_origin_dirty_style_sheet` flag to
ServiceWorkerNetworkProviderForFrame::GetSubresourceLoaderFactory().
So this CL adds these three arguments to CreateURLLoader() methods.

[1]: https://crrev.com/c/994152
[2]: https://crrev.com/c/5113665
[3]: https://groups.google.com/a/chromium.org/forum/#!topic/platform-architecture-dev/ntSzerId4gw

Bug: 1379780
Change-Id: I6e0a987d779b1fb428a4b27b813b3f53f2d1d781
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5121813
Reviewed-by: Erik Chen <erikchen@chromium.org>
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Reviewed-by: danakj <danakj@chromium.org>
Reviewed-by: Sean Topping <seantopping@chromium.org>
Reviewed-by: Richard (Torne) Coles <torne@chromium.org>
Commit-Queue: Tsuyoshi Horo <horo@chromium.org>
Reviewed-by: David Song <wintermelons@google.com>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1245656}
NOKEYCHECK=True
GitOrigin-RevId: 2455e700019a2766db2b51c19e9ec4616e62b509