[BreakoutBox] Reject invalid audio data in MediaStreamTrackProcessor
This CL causes writes of audio data to
MediaStreamAudioTrackUnderlyingSink (used by MSTP for audio)
with invalid audio parameters to be rejected. This type of data cannot
be handled by the audio stack downstream and leads to crashes.

MSATUS already rejects other types of invalid data, so this behavior
is consistent.

For extra security, this CL also adds a CHECK in
PushableMediaStreamAudioSource (downstream of MSATUS) to assert that
incoming audio samples are valid.

Fixed: 1513516
Change-Id: I3b419773413d45ff329d796723d7f84247776ed7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5182956
Commit-Queue: Guido Urdaneta <guidou@chromium.org>
Reviewed-by: Thomas Guilbert <tguilbert@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1245861}
NOKEYCHECK=True
GitOrigin-RevId: 9a19cb1cbf21980b7f55c92fc45e2c213ecb1b23