Rename runtime enabled feature for top-level 3PC deprecation trial
This CL renames the BREF for the top-level deprecation trial for
third-party cookies from "Tpcd1p" to "TopLevelTpcd". The new name is
more descriptive and less likely to be easily confused with the
counterpart that is for the third-aprty deprecation trial. The feature
is not yet used in code, so this is a no-op change.

Bug: 1513317
Change-Id: I2e76d686494a658c750b54856454e60bb9f6e3ff
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5186978
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Auto-Submit: Joshua Hood <jdh@chromium.org>
Commit-Queue: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1246112}
NOKEYCHECK=True
GitOrigin-RevId: 9af3edc2e28a9125a7a2650f040bd3f5c36e9e1e