Clear last element under mouse from the Event Handler.
When scrollbars get destroyed by PLSA::DestroyScrollbar the EventManager
class might still have a reference to the destroyed scrollbar. This
CL makes sure the reference is updated when the scrollbar is destroyed
so that no unintended calls to the destroyed scrollbar happen.

This was discovered when fixing tests that break when Fluent scrollbars
are enabled. The event manager was calling the Fluent scrollbar theme,
which checked the CSS property of a destroyed scrollbar and crashed in
the web test `scrollbars/custom-scrollbar-drag.html`

More details in the linked bug.

Bug: 1517340
Change-Id: I7293c1c90f90215eeb1e215962ff777fe001fd53
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5181885
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Reviewed-by: Yaroslav Shalivskyy <yshalivskyy@microsoft.com>
Commit-Queue: Gaston Rodriguez <gastonr@microsoft.com>
Reviewed-by: Olga Gerchikov <gerchiko@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1246115}
NOKEYCHECK=True
GitOrigin-RevId: 556a639302edb38380da70a49cd31bff92658b06