Accessibility: Add extension API to install FaceGaze assets
This change adds a private extension API that allows clients to install
the FaceGaze assets DLC. Once the DLC is downloaded, a callback will
be run with the DLC file bytes. In a follow-up CL, the FaceGaze
extension will use this API to download the DLC and use the bytes to
instantiate the mediapipe FaceLandmarker API, which can be used to
detect the location of facial landmarks and facial gestures.

Bug: b:309121742
Change-Id: I36b3ae6e1d800e1f34394bed8082e42068988850
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5169428
Reviewed-by: Katie Dektar <katie@chromium.org>
Commit-Queue: Akihiro Ota <akihiroota@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Steven Holte <holte@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1246572}
NOKEYCHECK=True
GitOrigin-RevId: fa7a0ed0639765ef09262e5508e1ed38cd7d17ea