[gc] Move Font objects in tests to being on-stack.
This is a preparation patch for moving FontFallbackList to being GC'd.

Font will become a GC'd part-object in the subsequent patch, and as
such non-GC'd objects (the test classes) cannot have them as part
objects. Move to being on the stack instead.

Bug: 1517025
Change-Id: Iab14407453c158728dfc6505a36572756b374a9a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5190825
Commit-Queue: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: Dominik Röttsches <drott@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1246604}
NOKEYCHECK=True
GitOrigin-RevId: 85950bc24d63f5ca146c1e6707f76ea02fba2c31