DOM: Generalize Observable "internal observer" handlers
In the Observable specification, recently
https://github.com/WICG/observable/pull/95 generalized the collection of
"next", "error", "complete" handlers into an "internal observer" concept
that the `Subscriber` interface is responsible for holding. This was an
abstraction exercise to make it possible to supply arbitrary observer
handlers, instead of just JavaScript callback functions passed into
`Observable::subscribe()`, via the `ObserverUnion`.

This abstraction is crucial for the Observable API, because it is not
only developer JavaScript code that is responsible for subscribing to
observables. Native "internal" code can also subscribe to an existing
Observable, via the Promise-returning operators [1]. These operators
internally subscribe to an Observable and return a Promise that resolves
or rejects according to the operator's semantics. The Promise resolves
or rejects with values accumulated via the "internal" next, error, and
complete handlers.

To achieve this in the implementation, we introduce a
`blink::ObservableInternalObserver` class mirroring the spec's "internal
observer" struct [2]. This is an abstract interface with overridable
`Next`, `Error`, and `Complete` algorithms. `blink::Subscriber` holds
onto this class, and subclasses of this interface provide the concrete
implementations of those algorithms.

Because no Promise-returning operators are implemented yet, the only
consumer/subclass of this abstract interface is the
`ScriptCallbackInternalObserver` that `observable.cc` provides. It is
responsible for implementing the logic that much of `blink::Subscriber`
currently implements (before this CL). This is typically trivial logic
like storing the user-supplied `ObserverUnion` callback functions,
calling them if they are non-null, and reporting exceptions if the
callbacks throw any errors.

With this logic outsourced to the specific
`ScriptCallbackInternalObserver`, and no longer hard-coded inside
`Subscriber`, Subscriber can be more concerned with logic pertaining to
subscription management, and less with Observer-specific logic.
Subsequent CLs will implement specific subclasses of this generic
interface to provide the Promise-returning operator implementations.

[1]: https://wicg.github.io/observable/#promise-returning-operators
[2]: https://wicg.github.io/observable/#internal-observer

R=masonf@chromium.org

Bug: 1485981
Change-Id: I524f0de89dab23a896b1f6f4783c0168b1fa6c0a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5178410
Commit-Queue: Dominic Farolino <dom@chromium.org>
Reviewed-by: Mason Freed <masonf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1246676}
NOKEYCHECK=True
GitOrigin-RevId: a4c0199ce9de31e56465f105372babbebd30a4b1