[inspector] Add 'charset' to Network.Response
MIME type and charset are parsed out of the Content-Type header so
technically the information is already available in the
"responseReceived" event.

Since MIME type is already a separate field and charset is crucial to
decode text resources that are sent in binary form (e.g. SVG) we also
include it in the "responseReceived" events. This also saves clients
from implementing a Content-Type parser.

Bug: 1509336
Change-Id: Ieac4a126aab62f975bdce4b5ae886abe3f96d63a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5134651
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1247044}
NOKEYCHECK=True
GitOrigin-RevId: 61ea84b3f9373853f3ccd147024ee4bf69ea5e47