Make IdentityUrlLoaderThrottle work on non-main thread
Currently IdentityUrlLoaderThrottle is not created by
URLLoaderThrottleProvider when CreateThrottles() is called on non-main
thread. So IdentityUrlLoaderThrottle don't work for subresource requests
from workers. This is not a big issue since login and logout requests
are rarely sent by workers.

But it will be a problen when we will call CreateThrottles() of a cloned
URLLoaderThrottleProvider on the background thread for normal
subresource requests from frames for BackgroundResourceFetch feature
(crrev.com/c/5113665).

So this CL does followings:
1. Keep the main thread task runner in providers
   (URLLoaderThrottleProviderImpl and
   ShellContentRendererUrlLoaderThrottleProvider).
2. Make CreateThrottles() method of the providers create
   IdentityUrlLoaderThrottle even when it is called on non-main thread.
3. In the callback of MaybeCreateIdentityUrlLoaderThrottle, use the
   main thread task runner to call blink::SetIdpSigninStatus() in the
   main thread when the callback is called on non-main thread.

Low-Coverage-Reason: COVERAGE_UNDERREPORTED
Bug: 1379780
Change-Id: I7c26ec390a10bd6732d7a8b9d0931efaa9f81bc7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5191809
Reviewed-by: danakj <danakj@chromium.org>
Reviewed-by: Erik Chen <erikchen@chromium.org>
Commit-Queue: Tsuyoshi Horo <horo@chromium.org>
Reviewed-by: Christian Biesinger <cbiesinger@chromium.org>
Reviewed-by: Antonio Sartori <antoniosartori@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1247051}
NOKEYCHECK=True
GitOrigin-RevId: f794e224cb310e911e164bd42d5e06ebbe13704f