Remove intersection ratio calculation
We were previously calculating the current intersection ratio to
call EnqueueEnteredViewport early (i.e. before we even create an
intersection observer). The calculation however was incorrect;
ratio_visible_area is the ratio of the visible area of the link with
the viewport area, and is not the ratio of the visible area of the
link with the total area of the link (which is what we actually want
to compute here).

Fortunately, the intersection observer created will trigger a call to
EnqueueEnteredViewport shortly after (the current check is done in
DidFinishLifecycleObserver; and the intersection observer would send a
notification in RunPostLifecycleSteps, which runs after the lifecycle
update synchronously), so we could just remove this check. Looking at
the history of this check [1], it was added to avoid making the
intersection observer observe the anchor entirely, but it looks like
we now always observe the anchor regardless of whether it is already
in the viewport or not.

[1] crrev.com/c/2922624

Change-Id: Ied9794ce7e8f509ab34104bb07c5097fb7a96cc0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5193145
Commit-Queue: Adithya Srinivasan <adithyas@chromium.org>
Reviewed-by: Jeremy Roman <jbroman@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1247290}
NOKEYCHECK=True
GitOrigin-RevId: b61f3715347d152210134ae5387bfc5fa8c05352