[bfcache] Add wpt for parser abort and bfcache
This CL adds a WPT that aborts parser in the middle of loading and tests if the page is restored from bfcache, and checks the reason.

Due to crbug.com/1517496 , I needed to add a failing expectation for linux-chrome. This is now being investigated and will be removed in the future.
Bug:1517496, 1518720

Change-Id: Ia20c5d1f673ffc69fc71c76e1f5c3022dd5c80cb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5109299
Reviewed-by: Mingyu Lei <leimy@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1247395}
NOKEYCHECK=True
GitOrigin-RevId: b465a450cc8d71993f5beb070237427dfd04851b