Make legacy colors store [0,255] internally
See discussion here: https://github.com/w3c/csswg-drafts/issues/9759
This CL fixes WPTs that are being were added here:
https://github.com/web-platform-tests/wpt/pull/43931/files

We were previously normalizing legacy srgb colors so that they could
be treated interchangeably with non-legacy srgb colors. Unfortunately
this breaks math for relative colors, for example:

rgb(from red calc(r - 127) g b)

Should result in a half red, not black.

The clearest way to fix this is simply to store legacy srgb colors in
the [0,255] range, so that we're not faking it on serialization
either. This of course makes conversions more complex as we have to
tread SRGBLegacy and SRGB as different color spaces, though they are
converted between one another by a simple multiplication or division
by 255.

Also in this CL:
  - Renamed ClampInt to ClampInt255, since that's what it does.
  - Added a SerializeInternal() function to blink::Color that both CSS,
Legacy and Canvas serializers all use. It's useful for debugging since
it does not do things like converting hsl/hwb to srgb.
  - Removed color-valid-relative-color.html from failure expectations,
so now we have an expectations file.

The six new failures in color-computed-relative-color are due to bad
tests that assume this mapping specifically does not happen.

Of the 42 failures in color-valid-relative-color, 36 are old and were
being suppressed by testExpectations. The 6 new ones are the same as
for the computed file.

Bug: 1517349
Change-Id: I5e4bb8a4d469e0d018ce5022accdb6bc22b4c4dc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5184777
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Aaron Krajeski <aaronhk@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1247537}
NOKEYCHECK=True
GitOrigin-RevId: b98dc5706a670bb7b2ba8b25a6e0e787e58d5aec