Remove non-streaming (deprecated) declarative shadow dom
The deprecation was completed on January 8, 2024, fully disabled
for M120 stable users. And the flag was disabled in code as of
121.0.6143.0, in November, 2023.

This CL removes all of the code supporting the deprecated/removed
version of the `shadowroot` attribute.

Bug: 1396384
Change-Id: Icbe4b264c2cb4b4425898310fca7c5f804cb930c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5187471
Auto-Submit: Mason Freed <masonf@chromium.org>
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Commit-Queue: Chris Harrelson <chrishtr@chromium.org>
Reviewed-by: David Baron <dbaron@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1247640}
NOKEYCHECK=True
GitOrigin-RevId: 5f4c1fc06a1a30c51d97b23146cc72004c89692e