Remove empty closed paths
As discussed in the github issue,
https://github.com/whatwg/html/issues/1079, it's generally agreed to
skip drawing line caps for closed paths. This is a follow up cl of:
https://chromium-review.googlesource.com/c/chromium/src/+/4864061.

In this cl, it verifies if each closed path is empty, then it removes
the empty closed subpaths from path. As the subpaths are removed, the
corresponding line caps are skipped.

Bug: 644067

Change-Id: I14c5c4819a31eb9249d01574ff1b5c60826c3b9e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4916516
Commit-Queue: Yi Xu <yiyix@google.com>
Reviewed-by: Jean-Philippe Gravel <jpgravel@chromium.org>
Reviewed-by: Justin Novosad <junov@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1247735}
NOKEYCHECK=True
GitOrigin-RevId: 572074cb06425797e7e110511db405134cf67e2f